requirements.txt
# Requirements for Enhanced RAG System v7.0
# ==========================================

## Core Dependencies
PyMuPDF>=1.23.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pydantic>=2.0.0
numpy>=1.24.0
aiohttp>=3.9.0
requests>=2.31.0

## Environment Configuration
python-dotenv>=1.0.0

## NEW in v7: Clipboard Support
pyperclip>=1.8.2

## Optional but Recommended
# For better progress display
tqdm>=4.66.0

# For colored CLI output
rich>=13.0.0

## Installation Instructions
# ==========================

# Basic installation:
# pip install -r requirements.txt

# For clipboard support on Linux, you may also need:
# Ubuntu/Debian: sudo apt-get install xclip xsel
# Fedora: sudo dnf install xclip xsel

## Environment Setup
# ===================

# 1. Copy .env.example to .env:
# cp .env.example .env

# 2. Edit .env and add your API keys

# 3. Or run the setup script:
# python setup.py
